PDF TO SPEECH
Which utilizes the PyTTsX3 and PyPDF2 modules, is a PDF to speech conversion tool. This model enables users to dictate PDF files and save them as audio files. With this feature, individuals can easily convert written documents into an audible format for convenience and accessibility purposes. This tool is perfect for people with visual impairments, those who prefer to listen instead of read, or individuals who want to multitask while reviewing documentsf